requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.7.2
streamlit==1.28.0
